DEV update victory
there's a strange memory leak, it may be associated with some of the event structure... customizing the tooltip seemed to bring it forward and highlight it in firefox. upgrading the npm appears to have helped, slightly readjusted the tooltip since victory appears to compensate for the voroni padding now. stack charts now appear to fill in the blanks